WIP Try to be able to maange many RTP conns
The idea is to be able t add connections, and each connection has its 
own state. When data is received on exported socks, it knows how to pass 
the data to the object per connection.

This allows for avahi to connect to all exported services, and then just 
send the midi data on the alsaseq.