Return proper IP at A queries
Actually returns the last it was contacted to, which is quite 
convoluted, but as it is single threaded it will be the proper answer to 
A queries.

It needs to parse first all the local addresses (so a route change will 
disrupt the service), and then match the remote address to the network 
it is connected to.

There is a provision for a defaut address, but it should actually never 
happen, as rptmidid is a local service only (at least the mDNS part).