Working connect network to something
Before the events were not dispatched to alsa.

Now the are dispatched to all connected elements, independant of the 
remote side name.

For example if it connects to synths A and B, it will export both names, 
but independently of where it ahs been connected it will finally go to 
both synths.