Service runs in realtime
Actually the rest of the server is not optimized as
much as it could be: mainly there are allocationsand
blocking points all around. But looks stable enough.
A lot of effort is needed to avoid all unnnecesary
blocks: debug logging, send and receive packets,
wait for data, and allocations. There are lists and
vectors all around, and new connectiosn are created
and destroyed on the same thread.

But preactically, it works wonderfully.