Pack messages and send at once
If messages are close in time (less than
half latency), and are of some type (note
on, offf or cc), store them in a vector, and
after the half latency of the first message
they all are sent together.

The idea is to prevent network traffic, and
as we sent them together, on the other
end they will be scheduled at the right
time.

If a non packable message comes, then
whatever was waiting is sent, and the new
message too.

But so far the start time of the packet is
not set right yet for this situation, nor
receiver waits to schedule midi messages.