Play notes with proper latency compensation
And also compensates the latency to avoid missing too many spots.

It totally prefers to have more latency than be jittery.

Works perfect on cable connections, but wifi is too jiterry, and finally latenc spikes.

Also latency is averaged to avoid go down
too fast when doing the latency check,
which tends to lower the latency.