Fix access after free on signal connection
The signal had two slots connected, but the first removed
itself because it was a disconneciton. This made the next
iterator to be invalid (ctually by implementation I think it
was calling again the same iterator as was moved to the
end of the internal memory representation.. but was invalid).

Anyway, now before any call it has to double chck is still a
valid call, all that doing a copy of the initial call list.

Less efficient, but safe.