Mved arg parsing to argv.cpp.
Also added support for do not announce all connecitons to ALSA /
Network.  This method only works if there is no firewall, as it
creates random open ports, and I dont know how to avoid this and
make them unique. So can be disabled, and make easier to do not
make mistakes.